+++ modules/Product/Models/Location/LocationStock.php
@@ -1,0 +1,16 @@
+<?php
+
+
+namespace Modules\Product\Models\Location;
+
+
+use App\BaseModel;
+
+class LocationStock extends BaseModel
+{
+    protected $fillable = [
+        'location_id',
+        'product_id',
+    ];
+    protected $table = 'location_product_stocks';
+}