+++ modules/Product/Traits/Location/HasLocationStock.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace Modules\Product\Traits\Location;
+
+
+use Modules\Product\Models\Location\LocationStock;
+
+trait HasLocationStock
+{
+    public function location_stocks(){
+        return $this->hasMany(LocationStock::class,'product_id');
+    }
+}